[fix] rating 정수에서 실수로 자료형 수정 (2024.8.1 김도연)

+++ src/main/java/com/kibwa/tbti/DTO/LocalcreatorDTO.java
@@ -26,7 +26,7 @@ public class LocalcreatorDTO {
     private String website;
     private String addressCategory1;
     private String addressCategory2;
-    private long rating;
+    private double rating;
     private char hiddenCategory;
     private String business_hours;
